refactor: rename dashboard to list patients
