+++ front-end/src/pages/ListPatients/styles.js
@@ -3,7 +3,6 @@ import styled from 'styled-components';
 export const Container = styled.div`
   width: 100vw;
   height: 100vh;
-  background-color: lightblue;
   display: flex;
   flex-direction: column;
   align-items: center;
